# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PowerCode AI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Snippet {
  id: number;
  title: string;
  code: string;
  language: string;
  /** @nullable */
  description?: string | null;
  isFavorited: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface CreateSnippetBody {
  title: string;
  code: string;
  language: string;
  /** @nullable */
  description?: string | null;
}

export interface UpdateSnippetBody {
  title?: string;
  code?: string;
  language?: string;
  /** @nullable */
  description?: string | null;
  isFavorited?: boolean;
}

export interface LanguageCount {
  language: string;
  count: number;
}

export interface SnippetStats {
  total: number;
  byLanguage: LanguageCount[];
  favorites: number;
  recentCount: number;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateOpenaiConversationBody {
  title: string;
}

export interface SendOpenaiMessageBody {
  content: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface OpenaiError {
  error: string;
}