# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PowerCode AI API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all code snippets
 */
export const ListSnippetsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  code: zod.string(),
  language: zod.string(),
  description: zod.string().nullish(),
  isFavorited: zod.boolean(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const ListSnippetsResponse = zod.array(ListSnippetsResponseItem);

/**
 * @summary Create a new code snippet
 */
export const CreateSnippetBody = zod.object({
  title: zod.string(),
  code: zod.string(),
  language: zod.string(),
  description: zod.string().nullish(),
});

/**
 * @summary Get a snippet by ID
 */
export const GetSnippetParams = zod.object({
  id: zod.coerce.number(),
});

export const GetSnippetResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  code: zod.string(),
  language: zod.string(),
  description: zod.string().nullish(),
  isFavorited: zod.boolean(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Update a snippet
 */
export const UpdateSnippetParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateSnippetBody = zod.object({
  title: zod.string().optional(),
  code: zod.string().optional(),
  language: zod.string().optional(),
  description: zod.string().nullish(),
  isFavorited: zod.boolean().optional(),
});

export const UpdateSnippetResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  code: zod.string(),
  language: zod.string(),
  description: zod.string().nullish(),
  isFavorited: zod.boolean(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Delete a snippet
 */
export const DeleteSnippetParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get snippet statistics
 */
export const GetSnippetStatsResponse = zod.object({
  total: zod.number(),
  byLanguage: zod.array(
    zod.object({
      language: zod.string(),
      count: zod.number(),
    }),
  ),
  favorites: zod.number(),
  recentCount: zod.number(),
});

/**
 * @summary List all conversations
 */
export const ListOpenaiConversationsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListOpenaiConversationsResponse = zod.array(
  ListOpenaiConversationsResponseItem,
);

/**
 * @summary Create a new conversation
 */
export const CreateOpenaiConversationBody = zod.object({
  title: zod.string(),
});

/**
 * @summary Get conversation with messages
 */
export const GetOpenaiConversationParams = zod.object({
  id: zod.coerce.number(),
});

export const GetOpenaiConversationResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.coerce.date(),
  messages: zod.array(
    zod.object({
      id: zod.number(),
      conversationId: zod.number(),
      role: zod.string(),
      content: zod.string(),
      createdAt: zod.coerce.date(),
    }),
  ),
});

/**
 * @summary Delete a conversation
 */
export const DeleteOpenaiConversationParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List messages in a conversation
 */
export const ListOpenaiMessagesParams = zod.object({
  id: zod.coerce.number(),
});

export const ListOpenaiMessagesResponseItem = zod.object({
  id: zod.number(),
  conversationId: zod.number(),
  role: zod.string(),
  content: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListOpenaiMessagesResponse = zod.array(
  ListOpenaiMessagesResponseItem,
);

/**
 * @summary Send a text message and receive a streaming text response
 */
export const SendOpenaiMessageParams = zod.object({
  id: zod.coerce.number(),
});

export const SendOpenaiMessageBody = zod.object({
  content: zod.string(),
});